feat: ActiveSpeakerFragment
