+++ app/src/main/java/live/hms/app2/ui/meeting/commons/VideoGridBaseFragment.kt
@@ -1,0 +1,300 @@
+package live.hms.app2.ui.meeting.commons
+
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.GridLayout
+import androidx.core.view.children
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.activityViewModels
+import live.hms.app2.databinding.GridItemVideoBinding
+import live.hms.app2.ui.meeting.MeetingTrack
+import live.hms.app2.ui.meeting.MeetingViewModel
+import live.hms.app2.ui.settings.SettingsStore
+import live.hms.app2.util.NameUtils
+import live.hms.app2.util.SurfaceViewRendererUtil
+import live.hms.app2.util.crashlyticsLog
+import live.hms.video.sdk.models.HMSSpeaker
+import org.webrtc.RendererCommon
+import kotlin.math.max
+import kotlin.math.min
+
+/**
+ * The Grid is created by building column by column.
+ * Example: For 4x2 (rows x columns)
+ *  - 3 videos will have 3 rows, 1 column
+ *  - 5 videos will have 4 rows, 2 columns
+ *  - 8 videos will have 4 rows, 2 columns
+ */
+abstract class VideoGridBaseFragment : Fragment() {
+  companion object {
+    private const val TAG = "VideoGridBase"
+  }
+
+  protected val settings: SettingsStore by lazy { SettingsStore(requireContext()) }
+  protected val meetingViewModel by activityViewModels<MeetingViewModel>()
+
+  // Determined using the onResume() and onPause()
+  private var isViewVisible = false
+
+  protected data class RenderedViewPair(
+    val binding: GridItemVideoBinding,
+    val video: MeetingTrack
+  )
+
+  private val bindedVideoTrackIds = mutableSetOf<String>()
+  protected val renderedViews = ArrayList<RenderedViewPair>()
+
+  protected val rows: Int
+    get() = min(max(1, renderedViews.size), settings.videoGridRows)
+
+  protected val columns: Int
+    get() {
+      val maxColumns = settings.videoGridColumns
+      val result = max(1, (renderedViews.size + rows - 1) / rows)
+      if (result > maxColumns) {
+        val videos = renderedViews.map { it.video }
+        throw IllegalStateException(
+          "At most ${settings.videoGridRows * maxColumns} videos are allowed. Provided $videos"
+        )
+      }
+
+      return result
+    }
+
+  protected val maxItems: Int
+    get() = settings.videoGridRows * settings.videoGridColumns
+
+  private fun updateGridLayoutDimensions(layout: GridLayout) {
+    crashlyticsLog(TAG, "updateGridLayoutDimensions: ${rows}x${columns}")
+
+    var childIdx: Pair<Int, Int>? = null
+    var colIdx = 0
+    var rowIdx = 0
+
+    layout.apply {
+      crashlyticsLog(TAG, "Updating GridLayout.spec for ${children.count()} children")
+      for (child in children) {
+        childIdx = Pair(rowIdx, colIdx)
+
+        val params = child.layoutParams as GridLayout.LayoutParams
+        params.rowSpec = GridLayout.spec(rowIdx, 1, 1f)
+        params.columnSpec = GridLayout.spec(colIdx, 1, 1f)
+
+        if (colIdx + 1 == columns) {
+          rowIdx += 1
+          colIdx = 0
+        } else {
+          colIdx += 1
+        }
+      }
+
+      crashlyticsLog(TAG, "Changed GridLayout's children spec with bottom-right at $childIdx")
+
+      // Forces maxIndex to be recalculated when rowCount/columnCount is set
+      requestLayout()
+
+      rowCount = rows
+      columnCount = columns
+    }
+  }
+
+  private fun createVideoView(parent: ViewGroup): GridItemVideoBinding {
+    return GridItemVideoBinding.inflate(
+      LayoutInflater.from(requireContext()),
+      parent,
+      false
+    )
+  }
+
+  protected fun bindSurfaceView(binding: GridItemVideoBinding, item: MeetingTrack) {
+    if (item.video == null || item.video?.isMute == true) return
+
+    SurfaceViewRendererUtil.bind(binding.videoCard.surfaceView, item).let {
+      if (it) {
+        binding.videoCard.surfaceView.visibility = View.VISIBLE
+        bindedVideoTrackIds.add(item.video!!.trackId)
+      }
+    }
+  }
+
+  private fun bindVideo(binding: GridItemVideoBinding, item: MeetingTrack) {
+    // FIXME: Add a shared VM with activity scope to subscribe to events
+    // binding.container.setOnClickListener { viewModel.onVideoItemClick?.invoke(item) }
+
+    binding.videoCard.apply {
+      name.text = item.peer.name
+      nameInitials.text = NameUtils.getInitials(item.peer.name)
+      iconScreenShare.visibility = if (item.isScreen) View.VISIBLE else View.GONE
+      iconAudioOff.visibility = if (
+        item.isScreen.not() &&
+        (item.audio == null || item.audio!!.isMute)
+      ) View.VISIBLE else View.GONE
+
+      if (item.video == null || item.video?.isMute == true) {
+        surfaceView.visibility = View.GONE
+      } else {
+        surfaceView.visibility = View.VISIBLE
+      }
+
+      // TODO: Add listener for video stream on/off -> Change visibility of surface renderer
+      surfaceView.apply {
+        setScalingType(RendererCommon.ScalingType.SCALE_ASPECT_BALANCED)
+        setEnableHardwareScaler(true)
+      }
+    }
+  }
+
+  private fun unbindSurfaceView(
+    binding: GridItemVideoBinding,
+    item: MeetingTrack,
+    metadata: String = ""
+  ) {
+    if (!bindedVideoTrackIds.contains(item.video?.trackId ?: "")) return
+
+    SurfaceViewRendererUtil.unbind(binding.videoCard.surfaceView, item, metadata).let {
+      if (it) {
+        binding.videoCard.surfaceView.visibility = View.INVISIBLE
+        bindedVideoTrackIds.remove(item.video!!.trackId)
+      }
+    }
+  }
+
+
+  protected fun updateVideos(layout: GridLayout, newVideos: Array<MeetingTrack?>) {
+    crashlyticsLog(
+      TAG,
+      "updateVideos(${newVideos.size}) -- presently ${renderedViews.size} items in grid"
+    )
+
+    val newRenderedViews = ArrayList<RenderedViewPair>()
+
+    // Remove all the views which are not required now
+    for (currentRenderedView in renderedViews) {
+      val newVideo = newVideos.find { it == currentRenderedView.video }
+      if (newVideo == null) {
+        crashlyticsLog(
+          TAG,
+          "updateVideos: Removing view for video=${currentRenderedView.video} in fragment=$tag"
+        )
+
+        layout.apply {
+          // Unbind only when view is visible to user
+          if (isViewVisible) unbindSurfaceView(
+            currentRenderedView.binding,
+            currentRenderedView.video
+          )
+          removeViewInLayout(currentRenderedView.binding.root)
+        }
+      }
+    }
+
+    for (_newVideo in newVideos) {
+      _newVideo?.also { newVideo ->
+
+        // Check if video already rendered
+        val renderedViewPair = renderedViews.find { it.video == newVideo }
+        if (renderedViewPair != null) {
+          crashlyticsLog(TAG, "updateVideos: Keeping view for video=$newVideo  in fragment=$tag")
+          newRenderedViews.add(renderedViewPair)
+
+          if (isViewVisible && !bindedVideoTrackIds.contains(newVideo.video?.trackId ?: "")) {
+            // This view is not yet initialized (possibly because when AudioTrack was added --
+            // VideoTrack was not present, hence had to create an empty tile)
+            bindSurfaceView(renderedViewPair.binding, newVideo)
+          }
+
+        } else {
+          crashlyticsLog(TAG, "updateVideos: Creating view for video=${newVideo} in fragment=$tag")
+
+          // Create a new view
+          val videoBinding = createVideoView(layout)
+          bindVideo(videoBinding, newVideo)
+
+          // Bind surfaceView when view is visible to user
+          if (isViewVisible) {
+            bindSurfaceView(videoBinding, newVideo)
+          }
+
+          layout.addView(videoBinding.root)
+          newRenderedViews.add(RenderedViewPair(videoBinding, newVideo))
+        }
+      }
+    }
+
+    crashlyticsLog(
+      TAG,
+      "updateVideos: Change grid items from ${renderedViews.size} -> ${newRenderedViews.size}"
+    )
+
+    renderedViews.clear()
+    renderedViews.addAll(newRenderedViews)
+
+    // Re-bind all the videos, this handles any changes made in isMute
+    for (view in renderedViews) {
+      bindVideo(view.binding, view.video)
+    }
+
+    updateGridLayoutDimensions(layout)
+  }
+
+  protected fun applySpeakerUpdates(speakers: Array<HMSSpeaker>) {
+    renderedViews.forEach { renderedView ->
+      val track = renderedView.video.audio
+      renderedView.binding.apply {
+        if (track == null || track.isMute) {
+          videoCard.audioLevel.apply {
+            text = "-"
+          }
+          container.strokeWidth = 0
+        } else {
+          val level = speakers.find { it.trackId == track.trackId }?.level ?: 0
+
+          videoCard.audioLevel.apply {
+            text = "$level"
+          }
+
+          when {
+            level >= 70 -> {
+              container.strokeWidth = 6
+            }
+            70 > level && level >= settings.silenceAudioLevelThreshold -> {
+              container.strokeWidth = 4
+            }
+            else -> {
+              container.strokeWidth = 0
+            }
+          }
+        }
+      }
+    }
+  }
+
+  override fun onResume() {
+    super.onResume()
+    crashlyticsLog(TAG, "Fragment=$tag onResume()")
+    isViewVisible = true
+
+    renderedViews.forEach {
+      bindSurfaceView(it.binding, it.video)
+    }
+  }
+
+  override fun onPause() {
+    super.onPause()
+    crashlyticsLog(TAG, "Fragment=$tag onPause()")
+    isViewVisible = false
+
+    renderedViews.forEach {
+      unbindSurfaceView(it.binding, it.video)
+    }
+  }
+
+  override fun onDestroy() {
+    super.onDestroy()
+    crashlyticsLog(TAG, "Fragment=$tag onDestroy()")
+
+    // Release all references to views
+    renderedViews.clear()
+  }
+}
